airflow/requirements.txt
psycopg2-binary==2.9.9
sqlalchemy>=1.4.28,<2.0
apache-airflow-providers-postgres==5.8.0
requests==2.31.0
